+++ saved-sites/saved-sites-impl/src/main/java/com/duckduckgo/savedsites/impl/folders/BookmarkFoldersViewModel.kt
@@ -16,7 +16,6 @@
 
 package com.duckduckgo.savedsites.impl.folders
 
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.duckduckgo.anvil.annotations.ContributesViewModel
@@ -26,13 +25,17 @@ import com.duckduckgo.savedsites.api.SavedSitesRepository
 import com.duckduckgo.savedsites.api.models.BookmarkFolder
 import com.duckduckgo.savedsites.api.models.BookmarkFolderItem
 import com.duckduckgo.savedsites.impl.dialogs.AddBookmarkFolderDialogFragment.AddBookmarkFolderListener
+import com.duckduckgo.savedsites.impl.edit.EditBookmarkViewModel.ViewState
 import com.duckduckgo.savedsites.impl.folders.BookmarkFoldersViewModel.Command.NewFolderCreatedUpdateTheStructure
 import com.duckduckgo.savedsites.impl.folders.BookmarkFoldersViewModel.Command.SelectFolder
 import javax.inject.Inject
 import kotlinx.coroutines.channels.BufferOverflow.DROP_OLDEST
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.receiveAsFlow
+import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import timber.log.Timber
 
@@ -51,13 +54,11 @@ class BookmarkFoldersViewModel @Inject constructor(
         object NewFolderCreatedUpdateTheStructure : Command()
     }
 
-    val viewState: MutableLiveData<ViewState> = MutableLiveData()
+    private val _viewState = MutableStateFlow(ViewState())
+    val viewState = _viewState.asStateFlow()
+
     private val command = Channel<Command>(1, DROP_OLDEST)
     fun commands(): Flow<Command> = command.receiveAsFlow()
-
-    init {
-        viewState.value = ViewState()
-    }
 
     fun fetchBookmarkFolders(
         selectedFolderId: String,
@@ -85,7 +86,9 @@ class BookmarkFoldersViewModel @Inject constructor(
     }
 
     private fun onFolderStructureCreated(folderStructure: List<BookmarkFolderItem>) {
-        viewState.postValue(viewState.value?.copy(folderStructure = folderStructure))
+        viewModelScope.launch(dispatcherProvider.main()) {
+            _viewState.update { ViewState(folderStructure = folderStructure) }
+        }
     }
 
     fun onItemSelected(bookmarkFolder: BookmarkFolder) {
